Moved date functions to their own modules.
